youtube-bigdata-agent/pyproject.toml
[project]
name = "youtube-channel-analyzer"
version = "0.1.0"
description = "A hybrid data pipeline and AI agent system to analyze YouTube channel trends."
authors = [
    { name = "Your Name", email = "your@email.com" },
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai[tools]>=0.30.0",
    "python-dotenv>=1.0.0",
    "google-api-python-client>=2.90.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "konlpy>=0.6.0",
]

[tool.uv.sources]
# This section is optional but can be useful
# pypi = { url = "https://pypi.org/simple" }

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.13.0",
    "wordcloud>=1.9.0",
]